[IMP] Split the objects in separate files as per OCA standard.
