+++ business_requirement_deliverable_default/models/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 # © 2016 Elico Corp (https://www.elico-corp.com).
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from . import business_requirement_deliverable_default
+from . import business_requirement
+from . import product
