+++ business_requirement_deliverable_default/models/product.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+# © 2016 Elico Corp (https://www.elico-corp.com).
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+from openerp import api, fields, models
+
+
+class ProductTemplate(models.Model):
+    _inherit = "product.template"
+
+    resource_lines = fields.One2many(
+        comodel_name='business.requirement.resource',
+        inverse_name='product_template_id',
+        string='Business Requirement Resources',
+        copy=True,
+    )
